third_party/blink/web_tests/external/wpt/svg/geometry/parsing/width-computed-expected.txt

This is a testharness.js-based test.
[FAIL] SVG Geometry Properties: getComputedStyle().width, <svg> initial
  assert_equals: initial expected "200px" but got "auto"
[FAIL] SVG Geometry Properties: getComputedStyle().width, <svg> presentation attribute
  assert_equals: presentation attribute expected "100px" but got "auto"
[FAIL] SVG Geometry Properties: getComputedStyle().width, <svg> inline style (auto)
  assert_equals: inline style (auto) expected "200px" but got "auto"
[FAIL] SVG Geometry Properties: getComputedStyle().width, <svg> inline style (percentage)
  assert_equals: inline style (percentage) expected "100px" but got "50%"
Harness: the test ran to completion.
